Refactor: Move ChromosomeSpec classes and update dependencies (#829)
