+++ src/main/java/com/verlumen/tradestream/backtesting/ChromosomeSpec.java
@@ -1,4 +1,4 @@
-package com.verlumen.tradestream.backtesting.params;
+package com.verlumen.tradestream.backtesting;
 
 import com.google.common.collect.Range;
 import io.jenetics.NumericChromosome;
