+++ src/main/java/com/verlumen/tradestream/backtesting/GAEngineFactoryImpl.java
@@ -1,7 +1,7 @@
 package com.verlumen.tradestream.backtesting;
 
 import com.google.inject.Inject;
-import com.verlumen.tradestream.backtesting.params.ChromosomeSpec;
+import com.verlumen.tradestream.backtesting.ChromosomeSpec;
 import com.verlumen.tradestream.backtesting.params.ParamConfig;
 import com.verlumen.tradestream.backtesting.params.ParamConfigManager;
 import io.jenetics.DoubleChromosome;
