+++ src/main/java/com/verlumen/tradestream/backtesting/params/BUILD
@@ -7,6 +7,7 @@ java_library(
     srcs = glob(["*.java"]),
     deps = [
         "//protos:strategies_java_proto",
+        "//src/main/java/com/verlumen/tradestream/backtesting:chromosome_spec",
         "//third_party:guava",
         "//third_party:jenetics",
         "//third_party:protobuf_java",
