+++ src/main/java/com/verlumen/tradestream/backtesting/params/ParamConfig.java
@@ -2,6 +2,7 @@ package com.verlumen.tradestream.backtesting.params;
 
 import com.google.common.collect.ImmutableList;
 import com.google.protobuf.Any;
+import com.verlumen.tradestream.backtesting.ChromosomeSpec;
 import com.verlumen.tradestream.strategies.StrategyType;
 import io.jenetics.Gene;
 import io.jenetics.NumericChromosome;
