+++ src/main/java/com/verlumen/tradestream/backtesting/params/SmaRsiParamConfig.java
@@ -2,6 +2,7 @@ package com.verlumen.tradestream.backtesting.params;
 
 import com.google.common.collect.ImmutableList;
 import com.google.protobuf.Any;
+import com.verlumen.tradestream.backtesting.ChromosomeSpec;
 import com.verlumen.tradestream.strategies.SmaRsiParameters;
 import com.verlumen.tradestream.strategies.StrategyType;
 import io.jenetics.IntegerChromosome;
