+++ src/test/java/com/verlumen/tradestream/backtesting/GenotypeConverterImplTest.java
@@ -11,7 +11,7 @@ import com.google.inject.testing.fieldbinder.Bind;
 import com.google.inject.testing.fieldbinder.BoundFieldModule;
 import com.google.protobuf.Any;
 import com.google.protobuf.Int32Value;
-import com.verlumen.tradestream.backtesting.params.ChromosomeSpec;
+import com.verlumen.tradestream.backtesting.ChromosomeSpec;
 import com.verlumen.tradestream.backtesting.params.ParamConfig;
 import com.verlumen.tradestream.backtesting.params.ParamConfigManager;
 import com.verlumen.tradestream.strategies.StrategyType;
