+++ src/test/java/com/verlumen/tradestream/backtesting/params/BUILD
@@ -6,6 +6,7 @@ java_test(
     srcs = ["SmaRsiParamConfigTest.java"],
     deps = [
         "//protos:strategies_java_proto",
+        "//src/main/java/com/verlumen/tradestream/backtesting:chromosome_spec",
         "//src/main/java/com/verlumen/tradestream/backtesting/params:params_lib",
         "//third_party:guava",
         "//third_party:guice",
