+++ src/test/java/com/verlumen/tradestream/backtesting/params/SmaRsiParamConfigTest.java
@@ -3,6 +3,7 @@ package com.verlumen.tradestream.backtesting.params;
 import static com.google.common.truth.Truth.assertThat;
 
 import com.google.common.collect.ImmutableList;
+import com.verlumen.tradestream.backtesting.ChromosomeSpec;
 import com.verlumen.tradestream.strategies.SmaRsiParameters;
 import io.jenetics.DoubleChromosome;
 import io.jenetics.IntegerChromosome;
